EAC-1814 Updated testcases
